requirements.txt
anthropic>=0.40
chromadb>=0.5
sentence-transformers>=3.0
langchain>=0.3
langchain-community>=0.3
pdfplumber>=0.11
streamlit>=1.40
ragas>=0.2
python-dotenv>=1.0
pytest>=8.0